+++ R/functions.R
@@ -122,6 +122,7 @@ predict_gam <- function(model, exclude_terms = NULL, length_out = 50, values = N
 #' @param exclude_terms Terms to be excluded from the prediction. Term names should be given as they appear in the model summary (for example, \code{"s(x0,x1)"}).
 #' @param split Columns to separate as a named list.
 #' @param sep Separator between columns (default is \code{"\\."}, which is the default with \code{}). If character, it is interpreted as a regular expression.
+#' @param transform Function used to transform the fitted values (useful for getting plots on the response scale).
 #'
 #' @examples
 #' library(mgcv)
@@ -132,8 +133,8 @@ predict_gam <- function(model, exclude_terms = NULL, length_out = 50, values = N
 #' pred <- get_gam_predictions(model, x2)
 #'
 #' @export
-get_gam_predictions <- function(model, series, series_length = 25, conditions = NULL, exclude_random = TRUE, exclude_terms = NULL, split = NULL, sep = "\\.", time_series) {
-    if(!missing(time_series)) {
+get_gam_predictions <- function(model, series, series_length = 25, conditions = NULL, exclude_random = TRUE, exclude_terms = NULL, split = NULL, sep = "\\.", time_series, transform = NULL) {
+    if (!missing(time_series)) {
       warning("This argument has been deprecated and will be removed in the future. Please use `series` instead.")
 
       series_q = dplyr::enquo(time_series)
@@ -251,6 +252,11 @@ get_gam_predictions <- function(model, series, series_length = 25, conditions = 
         exclude = exclude_these
     )
 
+    if (!is.null(transform)) {
+      trans_fun <- transform
+      predicted$fit <- trans_fun(predicted$fit)
+    }
+
     predicted_tbl <- cbind(fitted_series, predicted) %>%
         dplyr::mutate(
             CI_upper = fit + 1.96 * se.fit,
@@ -325,8 +331,8 @@ get_gam_predictions <- function(model, series, series_length = 25, conditions = 
 #' @importFrom rlang "quo_name"
 #' @importFrom stats "predict"
 #' @export
-plot_smooths <- function(model, series, comparison = NULL, facet_terms = NULL, conditions = NULL, exclude_random = TRUE, exclude_terms = NULL, series_length = 25, split = NULL, sep = "\\.", time_series) {
-    if(!missing(time_series)) {
+plot_smooths <- function(model, series, comparison = NULL, facet_terms = NULL, conditions = NULL, exclude_random = TRUE, exclude_terms = NULL, series_length = 25, split = NULL, sep = "\\.", transform = NULL, time_series) {
+    if (!missing(time_series)) {
       warning("This argument has been deprecated and will be removed in the future. Please use `series` instead.")
 
       series_q = dplyr::enquo(time_series)
@@ -345,7 +351,7 @@ plot_smooths <- function(model, series, comparison = NULL, facet_terms = NULL, c
     }
     outcome_q <- model$formula[[2]]
 
-    predicted_tbl <- get_gam_predictions(model, !!series_q, conditions, exclude_random = exclude_random, exclude_terms = exclude_terms, series_length = series_length, split = split, sep = sep)
+    predicted_tbl <- get_gam_predictions(model, !!series_q, conditions, exclude_random = exclude_random, exclude_terms = exclude_terms, series_length = series_length, split = split, sep = sep, transform = transform)
 
     smooths_plot <- predicted_tbl %>%
         ggplot2::ggplot(