+++ NEWS.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## Unreleased
+### Added
+- added `transform` argument in `get_gam_predictions()` and `plot_smooths()` (closes https://github.com/stefanocoretta/tidymv/issues/9)
+
 ### Changed
 - add option of setting values to `NULL` in `predict_gam()` when excluding terms to reduce computation time (also mentioned in the vignette)